requirements.txt
torch>=1.7.1
git+https://github.com/openai/CLIP.git
Pillow>=8.0.0
ftfy
regex
tqdm
